Y-28: extract utils functions
